Cut a few seconds from pipeline by entirely skipping image get after put
The previous behavior skips downloading the image after uploading it to ECR, but still takes some time - up to 25 seconds, based on a quick review of one pipeline - to execute the noop get. This change entirely removes the get, which is not necessary since the image is not used after the put step.